renaming and proper dir structure
