+++ src/main/java/com/springboot/app/springbootfirstapp/users/UserService.java
@@ -1,9 +1,11 @@
-package com.springboot.app.springbootfirstapp;
+package com.springboot.app.springbootfirstapp.users;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+
+import com.springboot.app.springbootfirstapp.DBUtil;
 
 public class UserService {
